qtwayland: Disable timer leak patch
The qtwayland patch being applied for Qt issue:

https://bugreports.qt.io/browse/QTBUG-79838

no longer applies with 5.15, and the qtwayland code has changed
enough that it is not clear if it is still required.  Disable
the patch until someone in a position to reproduce the issue can
determine if it is still a problem.

Bug-AGL: SPEC-3819

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: I21eaec04a1d07080936ae8cd22de82d0d4df54ec